+++ tests/ld/test_ld.py
@@ -10,7 +10,9 @@ from ontolutils import namespaces, urirefs, Thing
 from ontolutils.ex import dcat
 from ontolutils.namespacelib import M4I
 from ontolutils.namespacelib import SCHEMA
+from ontolutils.utils.qudt_units import qudt_lookup
 from rdflib import DCAT
+from rdflib.namespace import SKOS
 
 import h5rdmtoolbox as h5tbx
 from h5rdmtoolbox import __version__
@@ -902,3 +904,78 @@ WHERE {
             g1ttl = g1.serialize(format="ttl")
             g2ttl = g2.serialize(format="ttl")
             self.assertEqual(g1ttl, g2ttl)
+
+    def test_rdf_mappings(self):
+        with h5tbx.File(name="deleteme.hdf", mode="w") as h5:
+            ds = h5.create_dataset("u", data=[1, 2, 3])
+            ds.attrs["units"] = "m/s"
+            g = h5.create_group("grp")
+            g.attrs["long_name"] = "velocity"
+
+        def _get_qudt_unit(value):
+            return qudt_lookup.get(value, None)
+
+        rdf_mappings = {
+            "units": {
+                "predicate": M4I.hasUnit,
+                "object": _get_qudt_unit,
+            },
+            "long_name": {
+                "predicate": SKOS.altLabel
+            }
+        }
+        ttl = h5tbx.serialize(
+            h5.hdf_filename,
+            fmt="ttl",
+            rdf_mappings=rdf_mappings,
+            file_uri="https://example.org#"
+        )
+        self.assertEqual("""@prefix hdf: <http://purl.allotrope.org/ontologies/hdf5/1.8#> .
+@prefix m4i: <http://w3id.org/nfdi4ing/metadata4ing#> .
+@prefix skos: <http://www.w3.org/2004/02/skos/core#> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+<https://example.org#tmp0.hdf> a hdf:File ;
+    hdf:rootGroup <https://example.org#tmp0.hdf/> .
+
+hdf:H5T_INTEL_I64 a hdf:Datatype .
+
+<https://example.org#tmp0.hdf/> a hdf:Group ;
+    hdf:member <https://example.org#tmp0.hdf/grp>,
+        <https://example.org#tmp0.hdf/u> ;
+    hdf:name "/" .
+
+<https://example.org#tmp0.hdf/grp> a hdf:Group ;
+    hdf:attribute <https://example.org#tmp0.hdf/grp@long_name> ;
+    hdf:name "/grp" ;
+    skos:altLabel "velocity" .
+
+<https://example.org#tmp0.hdf/grp@long_name> a hdf:StringAttribute ;
+    hdf:data "velocity" ;
+    hdf:name "long_name" .
+
+<https://example.org#tmp0.hdf/u> a hdf:Dataset ;
+    hdf:attribute <https://example.org#tmp0.hdf/u@units> ;
+    hdf:dataspace <https://example.org#tmp0.hdf/u__dataspace> ;
+    hdf:datatype hdf:H5T_INTEGER,
+        hdf:H5T_INTEL_I64 ;
+    hdf:layout hdf:H5D_CONTIGUOUS ;
+    hdf:maximumSize 3 ;
+    hdf:name "/u" ;
+    hdf:rank 1 ;
+    hdf:size 3 ;
+    m4i:hasUnit <http://qudt.org/vocab/unit/M-PER-SEC> .
+
+<https://example.org#tmp0.hdf/u@units> a hdf:StringAttribute ;
+    hdf:data "m/s" ;
+    hdf:name "units" .
+
+<https://example.org#tmp0.hdf/u__dataspace> a hdf:SimpleDataspace ;
+    hdf:dimension <https://example.org#tmp0.hdf/u__dataspace_dimension_0> .
+
+<https://example.org#tmp0.hdf/u__dataspace_dimension_0> a hdf:DataspaceDimension ;
+    hdf:dimensionIndex 0 ;
+    hdf:size 3 .
+
+""".replace("tmp0.hdf", h5.hdf_filename.name), ttl)
+        h5.hdf_filename.unlink(missing_ok=True)